+++ lua/obsidian/client.lua
@@ -673,6 +673,68 @@ Client.resolve_link_async = function(self, link, callback)
   end)
 end
 
+--- Follow a link. If the link argument is `nil` we attempt to follow a link under the cursor.
+---
+---@param link string|?
+---@param opts { open_strategy: obsidian.config.OpenStrategy|? }|?
+Client.follow_link_async = function(self, link, opts)
+  opts = opts and opts or {}
+  self:resolve_link_async(link, function(res)
+    if res == nil then
+      return
+    end
+
+    if res.url ~= nil then
+      if self.opts.follow_url_func ~= nil then
+        self.opts.follow_url_func(res.url)
+      else
+        log.warn "This looks like a URL. You can customize the behavior of URLs with the 'follow_url_func' option."
+      end
+      return
+    end
+
+    local open_cmd = "e "
+    if opts.open_strategy ~= nil then
+      open_cmd = util.get_open_strategy(opts.open_strategy)
+    end
+
+    if res.note ~= nil then
+      -- Go to resolved note.
+      local path = assert(res.path)
+      return vim.schedule(function()
+        vim.api.nvim_command(open_cmd .. tostring(path))
+      end)
+    end
+
+    if res.link_type == search.RefTypes.Wiki or res.link_type == search.RefTypes.WikiWithAlias then
+      -- Prompt to create a new note.
+      return vim.schedule(function()
+        local confirmation = string.lower(vim.fn.input {
+          prompt = "Create new note '" .. res.location .. "'? [Y/n] ",
+        })
+        if confirmation == "" or confirmation == "y" or confirmation == "yes" then
+          -- Create a new note.
+          ---@type string|?, string[]
+          local id, aliases
+          if res.name == res.location then
+            aliases = {}
+          else
+            aliases = { res.name }
+            id = res.location
+          end
+
+          local note = self:new_note(res.name, id, nil, aliases)
+          vim.api.nvim_command(open_cmd .. tostring(note.path))
+        else
+          log.warn "Aborting"
+        end
+      end)
+    end
+
+    return log.err("Failed to resolve file '" .. res.location .. "'")
+  end)
+end
+
 --- Get the current note.
 ---
 ---@return obsidian.Note|?